Extract code related to the texlab invocation into separate module
